==> 01_create-empty-ncdf-files-ABUND.R <==
###########################################################################
####### Spatiotemporal decoupling of avian taxonomic and functional diversity, Jarzyna & Stagge
####### 01: Create empty netCDF files to hold species abundance data
###########################################################################

###########################################################################
## Load Packages
###########################################################################
require(ncdf4)
#install.packages("ebirdst")
require(ebirdst)
require(raster)
require(terra)
library(viridis)
library(sf)
library(rnaturalearth)

###########################################################################
## Check the Ebird data
###########################################################################
# (data that were posted in early 2021 are for 2019)
# (data for 2020 were not available for download at the time of analysis)
# (data used here is for 2019: 807 species)
dim(ebirdst_runs) #A full list of species and run names can be found in the ebirdst_runs dataframe; dimensions 610x12, there are 610 species
head(ebirdst_runs)
### Check the names of species and get the codes
spnames <- ebirdst_runs[,1:4]
spcodes <- spnames$species_code
### Set parameters to run through 
nweek <- 52
n_species <- length(spcodes)

###########################################################################
## Perform a test download to get the latitude longitude information
###########################################################################
# read in example to get x and y
sp.nam <- readRDS(file="Data/sp-abun-paths-2019.rds")
ab1 <- load_raster("abundance", path = sp.nam[4,5])
 
# cut ab 1 raster to desired dimensions
# Extract longitude
lon_seq <- seq(xmin(ab1), xmax(ab1), length.out=ab1@ncols) 
head(lon_seq)
# Extract latitude
lat_seq <- seq(ymax(ab1), ymin(ab1), length.out=ab1@nrows) 
head(lat_seq)
# Extract CRS
ab_crs <- crs(ab1)
ab_crs <-  ab_crs@projargs
ab_crs
# Time sequence
time_seq <- seq(1,52) - 1
# Species sequence
species_seq <- seq(1, n_species)

###########################################################################
## Create an empty NetCDF file to store species abundance estimates
###########################################################################
### Set filenames and location
#
for (i in 1:52){
nc_file <- file.path(paste0("Data/abundmean_allsp_week_",i,".nc"))

### Define dimensions
dim_lon <- ncdim_def( "x", units="m",  lon_seq, longname="eastward distance from southwest corner of domain in projection coordinates", create_dimvar=TRUE)
dim_lat <- ncdim_def( "y", units="m", lat_seq,  longname="northward distance from southwest corner of domain in projection coordinates", create_dimvar=TRUE)
dim_time <- ncdim_def( "time", "weeks since 2019-01-01", time_seq, unlim=TRUE)
dim_species <- ncdim_def( "species", "id", species_seq, create_dimvar=TRUE)

### Define new data variable
abund_var = ncvar_def("abundance", "unsure", list(dim_lat, dim_lon, dim_species), longname="Relative Abundance (unsure)",prec="double", missval= NA, compression = 6)
proj_def <- ncvar_def("sinu","1",NULL,NULL,longname="Sinusoidal",prec="char")

### Define file
nc <- nc_create(nc_file, list(abund_var, proj_def))

# put additional attributes into dimension and data variables
ncatt_put(nc,"x","axis","X")
ncatt_put(nc,"x","standard_name","projection_x_coordinate")
ncatt_put(nc,"x","_CoordinateAxisType","GeoX")
ncatt_put(nc,"y","axis","Y")
ncatt_put(nc,"y","standard_name","projection_y_coordinate")
ncatt_put(nc,"y","_CoordinateAxisType","GeoY")
ncatt_put(nc,"abundance","grid_mapping", "sinu")
ncatt_put(nc,"abundance","coordinates", "lat lon")

# put the CRS attributes
projname <- "Sinusoidal"
false_easting <- 0
false_northing <- 0
ncatt_put(nc,"sinu","name",projname)
ncatt_put(nc,"sinu","long_name",projname)
ncatt_put(nc,"sinu","grid_mapping_name",projname)
ncatt_put(nc,"sinu","longitude_of_central_meridian", 0)
ncatt_put(nc,"sinu","latitude_of_projection_origin",0)
#ncatt_put(nc,"sinu","standard_parallel", c(50.0, 50.0))
ncatt_put(nc,"sinu","false_easting",false_easting)
ncatt_put(nc,"sinu","false_northing",false_northing)
ncatt_put(nc,"sinu","_CoordinateTransformType","Projection")
ncatt_put(nc,"sinu","_CoordinateAxisTypes","GeoX GeoY")
ncatt_put(nc, 0, "Title", paste0("Week_",i,"_SpeciesMeanAbundance"))
ncatt_put(nc, 0, "Note1", "Generated by Marta Jarzyna")
cur_time = Sys.time()
ncatt_put(nc, 0, "Note2", paste("Modified", format(Sys.time(), "%Y%m%d_%H%M%S"), "on host", Sys.info()[4]))

nc_close(nc)
}
